Create LegacyIFrame component to use it in LegacyDashboardFrame and LegacySiteToolsFrame
