+++ ui/app/src/components/LegacySiteToolsFrame/LegacySiteToolsFrame.tsx
@@ -14,34 +14,18 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { IframeHTMLAttributes } from 'react';
-import { useEnv } from '../../utils/hooks';
-import { createStyles, makeStyles } from '@material-ui/core/styles';
+import React from 'react';
+import LegacyIFrame, { LegacyIFrameBaseProps } from '../LegacyIFrame/LegacyIFrame';
 
-interface LegacySiteToolsFrameProps {
+interface LegacySiteToolsFrameProps extends LegacyIFrameBaseProps {
   tool?: string;
   workAreaOnly?: boolean;
-  iframeProps?: IframeHTMLAttributes<any>;
 }
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    iframe: {
-      width: '100%',
-      height: '100%',
-      border: 'none'
-    }
-  })
-);
 
 function LegacySiteToolsFrame(props: LegacySiteToolsFrameProps) {
   const { tool, workAreaOnly = true, iframeProps } = props;
-  const authoringUrl = useEnv().authoringBase;
-  const classes = useStyles();
-  const iframeSrc = `${authoringUrl}/site-config${[workAreaOnly && '?mode=embedded', tool && `#tool/${tool}`]
-    .filter(Boolean)
-    .join('')}`;
-  return <iframe frameBorder={0} title="Site Tools" src={iframeSrc} className={classes.iframe} {...iframeProps} />;
+  const path = `/site-config${[workAreaOnly && '?mode=embedded', tool && `#tool/${tool}`].filter(Boolean).join('')}`;
+  return <LegacyIFrame path={path} title="Site Tools" iframeProps={iframeProps} />;
 }
 
 export default LegacySiteToolsFrame;
